TASK: Refactoring Part 1
