+++ ParkingOnBoard/Operation/ParkingOperation/ParkingOperationPark.cs
@@ -1,0 +1,9 @@
+﻿namespace ParkingOnBoard.Operation.ParkingOperation;
+
+public static class ParkingOperationPark
+{
+    public static void Park()
+    {
+
+    }
+}
